Massive package structure refactor, also move all modules configuration to Application.kt
