+++ src/main/kotlin/core/data/AdminData.kt
@@ -4,6 +4,7 @@ package dev.deadzone.core.data
  * Dummy data ready to use, typically for admin account.
  */
 object AdminData {
+    const val ADMIN_RESERVED_NAME = "givemeadmin"
     const val PLAYER_ID = "adm-id-123456"
     const val DISPLAY_NAME = "AdminPlayer" // also known as username or nickname
     const val EMAIL = "admin@admin.com"
