+++ src/main/kotlin/core/mission/model/LootParameter.kt
@@ -1,4 +1,4 @@
-package dev.deadzone.core.model
+package dev.deadzone.core.mission.model
 
 // Loot system config, contains variables that affects loot
 data class LootParameter(
