--- src/main/kotlin/module/HTTP.kt
@@ -1,21 +1,0 @@
-package dev.deadzone.module
-
-import io.ktor.http.*
-import io.ktor.server.application.*
-import io.ktor.server.plugins.cors.routing.*
-import io.ktor.server.plugins.statuspages.StatusPages
-import io.ktor.server.response.respondText
-
-fun Application.configureHTTP() {
-    install(CORS) {
-        allowHost("127.0.0.1:8080", schemes = listOf("http"))
-        allowHost("127.0.0.1:7777", schemes = listOf("http"))
-        allowHeader(HttpHeaders.ContentType)
-        allowMethod(HttpMethod.Get)
-    }
-    install(StatusPages) {
-        exception<Throwable> { call, cause ->
-            call.respondText(text = "500: $cause" , status = HttpStatusCode.InternalServerError)
-        }
-    }
-}
