--- src/main/kotlin/module/Websocket.kt
@@ -1,23 +1,0 @@
-package dev.deadzone.module
-
-import dev.deadzone.context.GlobalContext
-import io.ktor.server.application.Application
-import io.ktor.server.application.install
-import io.ktor.server.websocket.DefaultWebSocketServerSession
-import io.ktor.server.websocket.WebSockets
-import io.ktor.server.websocket.pingPeriod
-import io.ktor.server.websocket.timeout
-import io.ktor.websocket.Frame
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
-import java.util.concurrent.ConcurrentHashMap
-import kotlin.time.Duration.Companion.seconds
-
-fun Application.configureWebsocket() {
-    install(WebSockets) {
-        pingPeriod = 15.seconds
-        timeout = 15.seconds
-        masking = true
-    }
-}
