+++ src/main/kotlin/socket/handler/QuestProgress.kt
@@ -1,10 +1,10 @@
 package dev.deadzone.socket.handler
 
-import dev.deadzone.utils.PIOSerializer
-import dev.deadzone.socket.Connection
+import dev.deadzone.socket.protocol.PIOSerializer
+import dev.deadzone.socket.core.Connection
 import dev.deadzone.context.ServerContext
-import dev.deadzone.socket.utils.SocketMessage
-import dev.deadzone.socket.utils.SocketMessageHandler
+import dev.deadzone.socket.messaging.SocketMessage
+import dev.deadzone.socket.messaging.SocketMessageHandler
 
 /**
  * Handle `qp` message by:
