+++ src/main/kotlin/socket/handler/Save.kt
@@ -3,14 +3,14 @@ package dev.deadzone.socket.handler
 import dev.deadzone.core.compound.CompoundService
 import dev.deadzone.core.items.ItemFactory
 import dev.deadzone.core.mission.LootService
-import dev.deadzone.core.model.LootParameter
+import dev.deadzone.core.mission.model.LootParameter
 import dev.deadzone.core.model.game.data.*
 import dev.deadzone.core.survivor.SurvivorService
 import dev.deadzone.context.GlobalContext
 import dev.deadzone.module.LogConfigSocketError
 import dev.deadzone.module.LogConfigSocketToClient
 import dev.deadzone.module.Logger
-import dev.deadzone.socket.Connection
+import dev.deadzone.socket.core.Connection
 import dev.deadzone.context.ServerContext
 import dev.deadzone.socket.handler.saveresponse.compound.BuildingCreateBuyResponse
 import dev.deadzone.socket.handler.saveresponse.compound.BuildingCreateResponse
@@ -18,9 +18,9 @@ import dev.deadzone.socket.handler.saveresponse.compound.BuildingMoveResponse
 import dev.deadzone.socket.handler.saveresponse.crate.CrateUnlockResponse
 import dev.deadzone.socket.handler.saveresponse.mission.*
 import dev.deadzone.socket.handler.saveresponse.survivor.PlayerCustomResponse
-import dev.deadzone.socket.utils.SocketMessage
-import dev.deadzone.socket.utils.SocketMessageHandler
-import dev.deadzone.utils.PIOSerializer
+import dev.deadzone.socket.messaging.SocketMessage
+import dev.deadzone.socket.messaging.SocketMessageHandler
+import dev.deadzone.socket.protocol.PIOSerializer
 import io.ktor.util.date.*
 import java.util.*
 import kotlin.random.Random
