+++ src/main/kotlin/socket/handler/ZombieAttack.kt
@@ -1,10 +1,10 @@
 package dev.deadzone.socket.handler
 
-import dev.deadzone.utils.PIOSerializer
-import dev.deadzone.socket.Connection
+import dev.deadzone.socket.protocol.PIOSerializer
+import dev.deadzone.socket.core.Connection
 import dev.deadzone.context.ServerContext
-import dev.deadzone.socket.utils.SocketMessage
-import dev.deadzone.socket.utils.SocketMessageHandler
+import dev.deadzone.socket.messaging.SocketMessage
+import dev.deadzone.socket.messaging.SocketMessageHandler
 
 class ZombieAttackHandler(private val context: ServerContext): SocketMessageHandler {
     override fun match(message: SocketMessage): Boolean {
