+++ src/main/kotlin/socket/messaging/NetworkMessage.kt
@@ -1,0 +1,8 @@
+package dev.deadzone.socket.messaging
+
+/**
+ * Network message constants used by the game.
+ */
+object NetworkMessage {
+
+}
