+++ src/main/kotlin/socket/messaging/SaveDataMethod.kt
@@ -1,0 +1,8 @@
+package dev.deadzone.socket.messaging
+
+/**
+ * Save data method constants used by the game.
+ */
+object SaveDataMethod {
+
+}
