+++ src/main/kotlin/socket/tasks/ServerPushTaskDispatcher.kt
@@ -1,9 +1,9 @@
-package dev.deadzone.socket.utils
+package dev.deadzone.socket.tasks
 
 import dev.deadzone.module.LogConfigSocketError
 import dev.deadzone.module.LogSource
 import dev.deadzone.module.Logger
-import dev.deadzone.socket.Connection
+import dev.deadzone.socket.core.Connection
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
